# Switch plugin msgpack protocol to named format