--- a/crates/nu-plugin/src/serializers/msgpack.rs
+++ b/crates/nu-plugin/src/serializers/msgpack.rs
@@ -26,7 +26,7 @@ impl Encoder<PluginInput> for MsgPackSerializer {
         plugin_input: &PluginInput,
         writer: &mut impl std::io::Write,
     ) -> Result<(), nu_protocol::ShellError> {
-        rmp_serde::encode::write(writer, plugin_input).map_err(rmp_encode_err)
+        rmp_serde::encode::write_named(writer, plugin_input).map_err(rmp_encode_err)
     }
 
     fn decode(
@@ -46,7 +46,7 @@ impl Encoder<PluginOutput> for MsgPackSerializer {
         plugin_output: &PluginOutput,
         writer: &mut impl std::io::Write,
     ) -> Result<(), ShellError> {
-        rmp_serde::encode::write(writer, plugin_output).map_err(rmp_encode_err)
+        rmp_serde::encode::write_named(writer, plugin_output).map_err(rmp_encode_err)
     }
 
     fn decode(