@devyn 2024-04-19
Here's the benchmark comparison:

## Before

```
   ╰─ msgpack_encode                  │               │               │               │         │
      ├─ (100, 5)       65.46 µs      │ 71.57 µs      │ 65.56 µs      │ 65.79 µs      │ 100     │ 100
      ╰─ (10000, 15)    16.38 ms      │ 17.53 ms      │ 16.66 ms      │ 16.67 ms      │ 100     │ 100
```

## After

```
   ╰─ msgpack_encode                  │               │               │               │         │
      ├─ (100, 5)       95.15 µs      │ 102.4 µs      │ 95.64 µs      │ 96.46 µs      │ 100     │ 100
      ╰─ (10000, 15)    23.51 ms      │ 26.05 ms      │ 24.48 ms      │ 24.68 ms      │ 100     │ 100
```

So yeah, unfortunately, just as I expected, it's not small, around 50% worse. But I think it's worth it.