@fdncred 2024-04-19
Wow, that's a huge hit. We'll have to see how it plays out in reality. We need to keep this in our back pocket in case we need to increase the speed at some point.